+++ packages/rainbowkit/src/components/RainbowKitProvider/provideRainbowKitChains.ts
@@ -21,6 +21,8 @@ type ChainName =
   | 'goerli'
   | 'hardhat'
   | 'holesky'
+  | 'klaytn'
+  | 'klaytnBaobab'
   | 'kovan'
   | 'localhost'
   | 'mainnet'
@@ -36,6 +38,8 @@ type ChainName =
   | 'sepolia'
   | 'xdc'
   | 'xdcTestnet'
+  | 'zetachain'
+  | 'zetachainAthensTestnet'
   | 'zkSync'
   | 'zkSyncTestnet'
   | 'zora'
@@ -92,6 +96,11 @@ const hardhatIcon: IconMetadata = {
   iconUrl: async () => (await import('./chainIcons/hardhat.svg')).default,
 };
 
+const klaytnIcon: IconMetadata = {
+  iconBackground: 'transparent',
+  iconUrl: async () => (await import('./chainIcons/klaytn.svg')).default,
+};
+
 const optimismIcon: IconMetadata = {
   iconBackground: '#ff5a57',
   iconUrl: async () => (await import('./chainIcons/optimism.svg')).default,
@@ -105,6 +114,11 @@ const polygonIcon: IconMetadata = {
 const xdcIcon: IconMetadata = {
   iconBackground: '#f9f7ec',
   iconUrl: async () => (await import('./chainIcons/xdc.svg')).default,
+};
+
+const zetachainIcon: IconMetadata = {
+  iconBackground: '#000000',
+  iconUrl: async () => (await import('./chainIcons/zetachain.svg')).default,
 };
 
 const zkSyncIcon: IconMetadata = {
@@ -141,6 +155,8 @@ const chainMetadataByName: Record<ChainName, ChainMetadata | null> = {
   hardhat: { chainId: 31_337, ...hardhatIcon },
   holesky: { chainId: 17000, ...ethereumIcon },
   kovan: { chainId: 42, ...ethereumIcon },
+  klaytn: { chainId: 8_217, name: 'Klaytn', ...klaytnIcon },
+  klaytnBaobab: { chainId: 1_001, name: 'Klaytn Baobab', ...klaytnIcon },
   localhost: { chainId: 1_337, ...ethereumIcon },
   mainnet: { chainId: 1, name: 'Ethereum', ...ethereumIcon },
   optimism: { chainId: 10, name: 'Optimism', ...optimismIcon },
@@ -155,6 +171,12 @@ const chainMetadataByName: Record<ChainName, ChainMetadata | null> = {
   sepolia: { chainId: 11_155_111, ...ethereumIcon },
   xdc: { chainId: 50, name: 'XinFin', ...xdcIcon },
   xdcTestnet: { chainId: 51, ...xdcIcon },
+  zetachain: { chainId: 7000, name: 'ZetaChain', ...zetachainIcon },
+  zetachainAthensTestnet: {
+    chainId: 7001,
+    name: 'Zeta Athens',
+    ...zetachainIcon,
+  },
   zkSync: { chainId: 324, name: 'zkSync', ...zkSyncIcon },
   zkSyncTestnet: { chainId: 280, ...zkSyncIcon },
   zora: { chainId: 7777777, name: 'Zora', ...zoraIcon },