+++ packages/rainbowkit/src/config/getDefaultConfig.ts
@@ -1,7 +1,7 @@
-import { Transport } from 'viem';
+import type { Transport } from 'viem';
 import { http, CreateConfigParameters } from 'wagmi';
-import { WagmiProviderProps, createConfig } from 'wagmi';
-import { type RainbowKitChain } from '../components/RainbowKitProvider/RainbowKitChainContext';
+import { createConfig } from 'wagmi';
+import type { RainbowKitChain } from '../components/RainbowKitProvider/RainbowKitChainContext';
 import type { WalletList } from '../wallets/Wallet';
 import { computeWalletConnectMetaData } from '../wallets/computeWalletConnectMetaData';
 import { connectorsForWallets } from '../wallets/connectorsForWallets';
@@ -18,15 +18,16 @@ export type _chains = readonly [RainbowKitChain, ...RainbowKitChain[]];
 // It maps each 'Chain' id to a 'Transport'
 export type _transports = Record<_chains[number]['id'], Transport>;
 
-type WagmiConfigParameters = Omit<
-  CreateConfigParameters<_chains, _transports>,
-  // If you use 'client' you can't use 'transports' (we force to use 'transports')
-  // More info here https://wagmi.sh/core/api/createConfig#client
-  // We will also use our own 'connectors' instead of letting user specifying it
-  'client' | 'connectors'
->;
-
-interface GetDefaultConfigParameters extends WagmiConfigParameters {
+interface GetDefaultConfigParameters<
+  chains extends _chains,
+  transports extends _transports,
+> extends Omit<
+    CreateConfigParameters<chains, transports>,
+    // If you use 'client' you can't use 'transports' (we force to use 'transports')
+    // More info here https://wagmi.sh/core/api/createConfig#client
+    // We will also use our own 'connectors' instead of letting user specifying it
+    'client' | 'connectors'
+  > {
   appName: string;
   appDescription?: string;
   appUrl?: string;
@@ -35,17 +36,25 @@ interface GetDefaultConfigParameters extends WagmiConfigParameters {
   projectId: string;
 }
 
-const createDefaultTransports = (chains: _chains): _transports => {
-  const transportsObject = chains.reduce((acc: _transports, chain) => {
-    const key = chain.id as keyof _transports;
-    acc[key] = http() as _transports[keyof _transports]; // Type assertion here
+const createDefaultTransports = <
+  chains extends _chains,
+  transports extends _transports,
+>(
+  chains: chains,
+): transports => {
+  const transportsObject = chains.reduce((acc: transports, chain) => {
+    const key = chain.id as keyof transports;
+    acc[key] = http() as transports[keyof transports]; // Type assertion here
     return acc;
-  }, {} as _transports);
+  }, {} as transports);
 
   return transportsObject;
 };
 
-export const getDefaultConfig = ({
+export const getDefaultConfig = <
+  chains extends _chains,
+  transports extends _transports,
+>({
   appName,
   appDescription,
   appUrl,
@@ -53,8 +62,8 @@ export const getDefaultConfig = ({
   wallets,
   projectId,
   ...wagmiParameters
-}: GetDefaultConfigParameters): WagmiProviderProps['config'] => {
-  let { transports, chains, ...restWagmiParameters } = wagmiParameters;
+}: GetDefaultConfigParameters<chains, transports>) => {
+  const { transports, chains, ...restWagmiParameters } = wagmiParameters;
 
   const metadata = computeWalletConnectMetaData({
     appName,
@@ -85,14 +94,11 @@ export const getDefaultConfig = ({
     },
   );
 
-  if (!transports) {
-    transports = createDefaultTransports(chains);
-  }
-
   return createConfig({
     connectors,
     chains,
-    transports,
+    transports:
+      transports || createDefaultTransports<chains, transports>(chains),
     ...restWagmiParameters,
   });
 };